@@ -1,11 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="FluentValidation" Version="9.5.1" />
+    <PackageReference Include="Mapster" Version="7.4.0" />
     <PackageReference Include="PostSharp.Redist" Version="6.8.9" />
   </ItemGroup>
 
