@@ -3,18 +3,19 @@ using OvgBlog.DAL.Data;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace OvgBlog.Business.Abstract
 {
     public interface IContactService
     {
-        Task<IResult<Contact>> Create(Contact contact);
+        Task<IResult<Contact>> Create(Contact contact, CancellationToken cancellationToken);
 
-        Task<IResult<List<Contact>>> GetAll();
+        Task<IResult<List<Contact>>> GetAll(CancellationToken cancellationToken);
 
-        Task<IResult<Contact>> Get(Guid id);
+        Task<IResult<Contact>> Get(Guid id, CancellationToken cancellationToken);
 
-        Task<IResult<object>> Delete(Guid id);
+        Task<IResult<object>> Delete(Guid id, CancellationToken cancellationToken);
     }
 }
