@@ -5,27 +5,23 @@ using OvgBlog.DAL.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace OvgBlog.Business.Services
 {
-    public class CategoryService : ICategoryService
+    public class CategoryService(
+        IEntityRepository<Category> categoryRepository,
+        IEntityRepository<ArticleCategoryRelation> articleRelationRepository)
+        : ICategoryService
     {
-        readonly IEntityRepository<Category> _categoryRepository;
-        readonly IEntityRepository<ArticleCategoryRelation> _articleRelationRepository;
-        public CategoryService(IEntityRepository<Category> categoryRepository, IEntityRepository<ArticleCategoryRelation> articleRelationRepository)
-        {
-            _categoryRepository = categoryRepository;
-            _articleRelationRepository = articleRelationRepository;
-        }
-
-        public async Task<IResult<Category>> CategoryBySeoUrl(string seoUrl)
+        public async Task<IResult<Category>> CategoryBySeoUrl(string seoUrl, CancellationToken cancellationToken)
         {
             if (string.IsNullOrEmpty(seoUrl))
             {
                 return new Result<Category>(false, Message.CategoryNotFound);
             }
-            var category = await _categoryRepository.Get(x => x.SeoUrl == seoUrl);
+            var category = await categoryRepository.Get(cancellationToken, x => x.SeoUrl == seoUrl);
             if (category == null)
             {
                 return new Result<Category>(false, Message.CategoryNotFound);
@@ -34,54 +30,54 @@ namespace OvgBlog.Business.Services
         }
 
 
-        public async Task<IResult<Category>> Create(Category category)
+        public async Task<IResult<Category>> Create(Category category, CancellationToken cancellationToken)
         {
             if (category == null || string.IsNullOrEmpty(category.Name) || string.IsNullOrEmpty(category.SeoUrl))
             {
                 return new Result<Category>(false, Message.ModelNotValid);
             }
-            var result = await CategoryBySeoUrl(category.SeoUrl);
-            if (result.Success)
+            var result = await CategoryBySeoUrl(category.SeoUrl, cancellationToken);
+            if (result.IsSuccess)
             {
                 return new Result<Category>(false, Message.SeoUrlAlreadyTaken);
             }
             category.Id = Guid.NewGuid();
             category.CreatedDate = DateTime.Now;
-            await _categoryRepository.Create(category);
+            await categoryRepository.Create(category, cancellationToken);
             return new Result<Category>(true, category);
         }
 
-        public async Task<IResult<object>> Delete(Guid id)
+        public async Task<IResult<object>> Delete(Guid id, CancellationToken cancellationToken)
         {
             if (id == Guid.Empty)
             {
                 return new Result<object>(false, Message.CategoryNotFound);
             }
-            var categoryEntity = await _categoryRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var categoryEntity = await categoryRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (categoryEntity == null)
             {
                 return new Result<object>(false, Message.CategoryNotFound);
             }
-            var articleEntities = await _articleRelationRepository.GetAll(x => x.CategoryId == categoryEntity.Id && !x.IsDeleted);
+            var articleEntities = await articleRelationRepository.GetAll(cancellationToken, x => x.CategoryId == categoryEntity.Id && !x.IsDeleted);
             foreach (var item in articleEntities)
             {
                 item.IsDeleted = true;
                 item.DeletedDate = DateTime.Now;
-                await _articleRelationRepository.Update(item);
+                await articleRelationRepository.Update(item, cancellationToken);
             }
             categoryEntity.IsDeleted = true;
             categoryEntity.DeletedDate = DateTime.Now;
-            await _categoryRepository.Update(categoryEntity);
+            await categoryRepository.Update(categoryEntity, cancellationToken);
             return new Result<object>(true);
         }
 
-        public async Task<IResult<Category>> FindCategoryIdByName(string name)
+        public async Task<IResult<Category>> FindCategoryIdByName(string name, CancellationToken cancellationToken)
         {
             if (string.IsNullOrEmpty(name))
             {
                 return new Result<Category>(false, Message.NameNotFound);
             }
-            var category = await _categoryRepository.Get(x => x.Name == name && !x.IsDeleted);
+            var category = await categoryRepository.Get(cancellationToken, x => x.Name == name && !x.IsDeleted);
             if (category == null)
             {
                 return new Result<Category>(false, Message.CategoryNotFound);
@@ -89,19 +85,19 @@ namespace OvgBlog.Business.Services
             return new Result<Category>(true, category);
         }
 
-        public async Task<IResult<IEnumerable<Category>>> GetAll()
+        public async Task<IResult<IEnumerable<Category>>> GetAll(CancellationToken cancellationToken)
         {
-            var list = await _categoryRepository.GetAll(x => !x.IsDeleted);
+            var list = await categoryRepository.GetAll(cancellationToken, x => !x.IsDeleted);
             return new Result<IEnumerable<Category>>(true, list);
         }
 
-        public async Task<IResult<Category>> GetById(Guid id)
+        public async Task<IResult<Category>> GetById(Guid id, CancellationToken cancellationToken)
         {
             if (id == Guid.Empty)
             {
                 return new Result<Category>(false, Message.IdIsNotValid);
             }
-            var categoryEntity = await _categoryRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var categoryEntity = await categoryRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (categoryEntity == null)
             {
                 return new Result<Category>(false, Message.IdIsNotValid);
@@ -109,19 +105,19 @@ namespace OvgBlog.Business.Services
             return new Result<Category>(true, categoryEntity);
         }
 
-        public async Task<IResult<Category>> Update(Category category)
+        public async Task<IResult<Category>> Update(Category category, CancellationToken cancellationToken)
         {
             if (category == null || string.IsNullOrEmpty(category.Name) || string.IsNullOrEmpty(category.SeoUrl))
             {
                 return new Result<Category>(false, Message.ModelNotValid);
             }
-            var categoryEntity = await _categoryRepository.Get(x => x.Id == category.Id && !x.IsDeleted);
+            var categoryEntity = await categoryRepository.Get(cancellationToken, x => x.Id == category.Id && !x.IsDeleted);
             if (categoryEntity == null)
             {
                 return new Result<Category>(false, Message.CategoryNotFound);
             }
             category.UpdatedDate = DateTime.Now;
-            categoryEntity = await _categoryRepository.Update(category);
+            categoryEntity = await categoryRepository.Update(category, cancellationToken);
             return new Result<Category>(true, categoryEntity);
         }
     }
