@@ -5,6 +5,7 @@ using OvgBlog.DAL.Data;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace OvgBlog.Business.Services
@@ -18,58 +19,58 @@ namespace OvgBlog.Business.Services
             _commentRepository = commentRepository;
             _articleRepository = articleRepository;
         }
-        public async Task<IResult<Comment>> Create(Comment comment)
+        public async Task<IResult<Comment>> Create(Comment comment, CancellationToken cancellationToken)
         {
             if (comment == null || string.IsNullOrEmpty(comment.Name) || string.IsNullOrEmpty(comment.Body) || comment.ArticleId == Guid.Empty)
             {
                 return new Result<Comment>(false, Message.ModelNotValid);
             }
-            var articleEntity = _articleRepository.Get(x => x.Id == comment.ArticleId && !x.IsDeleted);
+            var articleEntity = _articleRepository.Get(cancellationToken, x => x.Id == comment.ArticleId && !x.IsDeleted);
             if (articleEntity == null)
             {
                 return new Result<Comment>(false, Message.CommentNotFound);
             }
             comment.Id = Guid.NewGuid();
             comment.CreatedDate = DateTime.Now;
-            await _commentRepository.Create(comment);
+            await _commentRepository.Create(comment, cancellationToken);
             return new Result<Comment>(true, comment);
         }
 
-        public async Task<IResult<object>> Delete(Guid id)
+        public async Task<IResult<object>> Delete(Guid id, CancellationToken cancellationToken)
         {
             if (id == Guid.Empty)
             {
                 return new Result<object>(false, Message.IdIsNotValid);
             }
-            var commentEntity = await _commentRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var commentEntity = await _commentRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (commentEntity == null)
             {
                 return new Result<object>(false, Message.CommentNotFound);
             }
-            var articleEntity = _articleRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var articleEntity = _articleRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (articleEntity == null)
             {
                 return new Result<object>(false, Message.CommentNotFound);
             }
             commentEntity.IsDeleted = true;
             commentEntity.DeletedDate = DateTime.Now;
-            await _commentRepository.Update(commentEntity);
+            await _commentRepository.Update(commentEntity, cancellationToken);
             return new Result<object>(true);
         }
 
-        public async Task<IResult<IEnumerable<Comment>>> GetAll()
+        public async Task<IResult<IEnumerable<Comment>>> GetAll(CancellationToken cancellationToken)
         {
-            var list = await _commentRepository.GetAll(x => !x.IsDeleted);
+            var list = await _commentRepository.GetAll(cancellationToken, x => !x.IsDeleted);
             return new Result<IEnumerable<Comment>>(true, list);
         }
 
-        public async Task<IResult<Comment>> GetById(Guid id)
+        public async Task<IResult<Comment>> GetById(Guid id, CancellationToken cancellationToken)
         {
             if (id == Guid.Empty)
             {
                 return new Result<Comment>(false, Message.IdIsNotValid);
             }
-            var commentEntity = await _commentRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var commentEntity = await _commentRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (commentEntity == null)
             {
                 return new Result<Comment>(false, Message.CommentNotFound);
@@ -77,24 +78,24 @@ namespace OvgBlog.Business.Services
             return new Result<Comment>(true, commentEntity);
         }
 
-        public async Task<IResult<Comment>> Update(Comment comment)
+        public async Task<IResult<Comment>> Update(Comment comment, CancellationToken cancellationToken)
         {
             if (comment == null || string.IsNullOrEmpty(comment.Name) || string.IsNullOrEmpty(comment.Body) || comment.ArticleId == Guid.Empty)
             {
                 return new Result<Comment>(false, Message.ModelNotValid);
             }
-            var commentEntity = await _commentRepository.Get(x => x.Id == comment.Id && !x.IsDeleted);
+            var commentEntity = await _commentRepository.Get(cancellationToken, x => x.Id == comment.Id && !x.IsDeleted);
             if (commentEntity == null)
             {
                 return new Result<Comment>(false, Message.CommentNotFound);
             }
-            var articleEntity = await _articleRepository.Get(x => x.Id == comment.ArticleId && !x.IsDeleted);
+            var articleEntity = await _articleRepository.Get(cancellationToken, x => x.Id == comment.ArticleId && !x.IsDeleted);
             if (articleEntity == null)
             {
                 return new Result<Comment>(false, Message.CommentNotFound);
             }
             comment.UpdatedDate = DateTime.Now;
-            commentEntity = await _commentRepository.Update(comment);
+            commentEntity = await _commentRepository.Update(comment, cancellationToken);
             return new Result<Comment>(true, commentEntity);
 
         }
