@@ -5,6 +5,7 @@ using OvgBlog.DAL.Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace OvgBlog.Business.Services
@@ -16,7 +17,7 @@ namespace OvgBlog.Business.Services
         {
             _userRepository = userRepository;
         }
-        public async Task<IResult<User>> Create(User user)
+        public async Task<IResult<User>> Create(User user, CancellationToken cancellationToken)
         {
             if (user == null || string.IsNullOrEmpty(user.Name) || string.IsNullOrEmpty(user.Password))
             {
@@ -24,35 +25,35 @@ namespace OvgBlog.Business.Services
             }
             user.Id = Guid.NewGuid();
             user.CreatedDate = DateTime.Now;
-            await _userRepository.Create(user);
+            await _userRepository.Create(user, cancellationToken);
             return new Result<User>(true, user);
 
         }
 
-        public async Task<IResult<object>> Delete(Guid id)
+        public async Task<IResult<object>> Delete(Guid id, CancellationToken cancellationToken)
         {
             if (id == Guid.Empty)
             {
                 return new Result<object>(false, Message.IdIsNotValid);
             }
-            var userEntity = await _userRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var userEntity = await _userRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (userEntity == null)
             {
                 return new Result<object>(false, Message.UserNotFound);
             }
             userEntity.IsDeleted = true;
             userEntity.DeletedDate = DateTime.Now;
-            await _userRepository.Update(userEntity);
+            await _userRepository.Update(userEntity, cancellationToken);
             return new Result<object>(true);
         }
 
-        public async Task<IResult<bool>> CheckUser(string userName, string password)
+        public async Task<IResult<bool>> CheckUser(string userName, string password, CancellationToken cancellationToken)
         {
             if (string.IsNullOrEmpty(userName) || string.IsNullOrEmpty(password))
             {
                 return new Result<bool>(false, Message.UserNotFound);
             }
-            var user = await _userRepository.Get(x => x.Name == userName || x.Email == userName);
+            var user = await _userRepository.Get(cancellationToken, x => x.Name == userName || x.Email == userName);
             if (user == null || user.Name == null || user.Password == null)
             {
                 return new Result<bool>(false, Message.UserNotFound);
@@ -64,13 +65,13 @@ namespace OvgBlog.Business.Services
             return new Result<bool>(true);
         }
 
-        public async Task<IResult<User>> GetById(Guid id)
+        public async Task<IResult<User>> GetById(Guid id, CancellationToken cancellationToken)
         {
             if (id == Guid.Empty)
             {
                 return new Result<User>(false, Message.IdIsNotValid);
             }
-            var userEntity = await _userRepository.Get(x => x.Id == id && !x.IsDeleted);
+            var userEntity = await _userRepository.Get(cancellationToken, x => x.Id == id && !x.IsDeleted);
             if (userEntity == null)
             {
                 return new Result<User>(false, Message.UserNotFound);
@@ -78,19 +79,19 @@ namespace OvgBlog.Business.Services
             return new Result<User>(true, userEntity);
         }
 
-        public async Task<IResult<User>> Update(User user)
+        public async Task<IResult<User>> Update(User user, CancellationToken cancellationToken)
         {
             if (user == null || string.IsNullOrEmpty(user.Name) || string.IsNullOrEmpty(user.Password))
             {
                 return new Result<User>(false, Message.ModelNotValid);
             }
-            var userEntity = await _userRepository.Get(x => x.Id == user.Id && !x.IsDeleted);
+            var userEntity = await _userRepository.Get(cancellationToken, x => x.Id == user.Id && !x.IsDeleted);
             if (userEntity == null)
             {
                 return new Result<User>(false, Message.UserNotFound);
             }
             user.UpdatedDate = DateTime.Now;
-            userEntity = await _userRepository.Update(user);
+            userEntity = await _userRepository.Update(user, cancellationToken);
             return new Result<User>(true, userEntity);
         }
     }
