@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using OvgBlog.DAL.Data.Base;
 
 #nullable disable
 
 namespace OvgBlog.DAL.Data
 {
-    public partial class Category
+    public class Category : BaseEntity
     {
         public Category()
         {
@@ -16,11 +17,6 @@ namespace OvgBlog.DAL.Data
         public string Name { get; set; }
         public string ImageUrl { get; set; }
         public string SeoUrl { get; set; }
-        public bool IsDeleted { get; set; }
-        public bool IsActive { get; set; }
-        public DateTime? CreatedDate { get; set; }
-        public DateTime? DeletedDate { get; set; }
-        public DateTime? UpdatedDate { get; set; }
 
         public virtual ICollection<ArticleCategoryRelation> ArticleCategoryRelations { get; set; }
     }
