@@ -1,17 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
+using OvgBlog.DAL.Data.Base;
 
 #nullable disable
 
 namespace OvgBlog.DAL.Data
 {
-    public partial class Contact
+    public class Contact : BaseEntity
     {
         public string Name { get; set; }
         public string Email { get; set; }
         public string Body { get; set; }
         public Guid Id { get; set; }
         public DateTime SendDate { get; set; }
-        public bool IsDeleted { get; set; }
     }
 }
