@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using OvgBlog.DAL.Data.Base;
 
 #nullable disable
 
 namespace OvgBlog.DAL.Data
 {
-    public partial class Tag
+    public class Tag : BaseEntity
     {
         public Tag()
         {
@@ -15,11 +16,6 @@ namespace OvgBlog.DAL.Data
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string SeoUrl { get; set; }
-        public bool IsDeleted { get; set; }
-        public bool IsActive { get; set; }
-        public DateTime? CreatedDate { get; set; }
-        public DateTime? DeletedDate { get; set; }
-        public DateTime? UpdatedDate { get; set; }
 
         public virtual ICollection<ArticleTagRelation> ArticleTagRelations { get; set; }
     }
