@@ -1,6 +1,9 @@
 ﻿using System;
+using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.Extensions.Configuration;
+using OvgBlog.DAL.Constant;
 using OvgBlog.DAL.Data;
 
 #nullable disable
@@ -10,11 +13,14 @@ namespace OvgBlog.DAL
     public partial class OvgBlogContext : DbContext
     {     
 
-        public OvgBlogContext(DbContextOptions<OvgBlogContext> options)
+        public OvgBlogContext(DbContextOptions<OvgBlogContext> options, IConfiguration configuration)
             : base(options)
         {
+            _configuration = configuration;
         }
 
+        private IConfiguration _configuration { get; }
+        
         public virtual DbSet<Article> Articles { get; set; }
         public virtual DbSet<ArticleCategoryRelation> ArticleCategoryRelations { get; set; }
         public virtual DbSet<ArticleTagRelation> ArticleTagRelations { get; set; }
@@ -26,17 +32,17 @@ namespace OvgBlog.DAL
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
-            modelBuilder.HasAnnotation("Relational:Collation", "Turkish_CI_AS");
-
+            base.OnModelCreating(modelBuilder);
+            
             modelBuilder.Entity<Article>(entity =>
             {
                 entity.ToTable("Article");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
                 entity.Property(e => e.ImageUrl).HasMaxLength(500);
 
@@ -48,7 +54,7 @@ namespace OvgBlog.DAL
                     .IsRequired()
                     .HasMaxLength(200);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
 
                 entity.HasOne(d => d.User)
                     .WithMany(p => p.Articles)
@@ -61,13 +67,13 @@ namespace OvgBlog.DAL
             {
                 entity.ToTable("ArticleCategoryRelation");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
 
                 entity.HasOne(d => d.Article)
                     .WithMany(p => p.ArticleCategoryRelations)
@@ -86,13 +92,13 @@ namespace OvgBlog.DAL
             {
                 entity.ToTable("ArticleTagRelation");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
 
                 entity.HasOne(d => d.Article)
                     .WithMany(p => p.ArticleTagRelations)
@@ -111,11 +117,11 @@ namespace OvgBlog.DAL
             {
                 entity.ToTable("Category");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
                 entity.Property(e => e.ImageUrl).HasMaxLength(250);
 
@@ -127,22 +133,22 @@ namespace OvgBlog.DAL
                     .IsRequired()
                     .HasMaxLength(50);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
             });
 
             modelBuilder.Entity<Comment>(entity =>
             {
                 entity.ToTable("Comment");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
                 entity.Property(e => e.Body)
                     .IsRequired()
                     .HasMaxLength(250);
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
                 entity.Property(e => e.Email).HasMaxLength(50);
 
@@ -150,7 +156,7 @@ namespace OvgBlog.DAL
                     .IsRequired()
                     .HasMaxLength(50);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
 
                 entity.HasOne(d => d.Article)
                     .WithMany(p => p.Comments)
@@ -163,7 +169,7 @@ namespace OvgBlog.DAL
             {
                 entity.ToTable("Contact");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
                 entity.Property(e => e.Body)
                     .IsRequired()
@@ -177,18 +183,18 @@ namespace OvgBlog.DAL
                     .IsRequired()
                     .HasMaxLength(50);
 
-                entity.Property(e => e.SendDate).HasColumnType("datetime");
+                entity.Property(e => e.SendDate);
             });
 
             modelBuilder.Entity<Tag>(entity =>
             {
                 entity.ToTable("Tag");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
                 entity.Property(e => e.Name)
                     .IsRequired()
@@ -198,18 +204,18 @@ namespace OvgBlog.DAL
                     .IsRequired()
                     .HasMaxLength(20);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
             });
 
             modelBuilder.Entity<User>(entity =>
             {
                 entity.ToTable("User");
 
-                entity.Property(e => e.Id).ValueGeneratedNever();
+                entity.Property(e => e.Id).ValueGeneratedOnAdd();
 
-                entity.Property(e => e.CreatedDate).HasColumnType("datetime");
+                entity.Property(e => e.CreatedDate);
 
-                entity.Property(e => e.DeletedDate).HasColumnType("datetime");
+                entity.Property(e => e.DeletedDate);
 
                 entity.Property(e => e.Email).HasMaxLength(50);
 
@@ -221,12 +227,26 @@ namespace OvgBlog.DAL
                     .IsRequired()
                     .HasMaxLength(50);
 
-                entity.Property(e => e.UpdatedDate).HasColumnType("datetime");
+                entity.Property(e => e.UpdatedDate);
             });
-
-            OnModelCreatingPartial(modelBuilder);
+        }
+        
+        public void Seed(IConfiguration configuration)
+        {
+            if (Users.Any(u => u.Email == configuration[UserConstants.AdminEmail])) return;
+            
+            Users.Add(new User
+            {
+                Id = Guid.NewGuid(),
+                Name = configuration[UserConstants.AdminUserName],
+                Email = configuration[UserConstants.AdminEmail],
+                Password = configuration[UserConstants.AdminPassword],
+                IsDeleted = false,
+                IsActive = true,
+                CreatedDate = DateTime.UtcNow
+            });
+            SaveChanges();
         }
 
-        partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
     }
 }
