@@ -30,13 +30,13 @@ namespace OvgBlog.UI.Pages
         {
             if (string.IsNullOrEmpty(UserName) || string.IsNullOrEmpty(Password))
             {
-                Message = "T�m alanlar� eksiksiz doldurun!";
+                Message = "T�m alanlar� eksiksiz doldurun!";
                 return Page();
             }
             var result = await _userService.CheckUser(UserName, Password);
-            if (!result.Success)
+            if (!result.IsSuccess)
             {
-                Message = "Ad�n�z ve ya �ifreniz yanl��t�r.";
+                Message = "Ad�n�z ve ya �ifreniz yanl��t�r.";
                 return Page();
             }
             var claims = new List<Claim> {
