@@ -30,7 +30,7 @@ namespace OvgBlog.UI
             Configuration = configuration;
         }
 
-        public IConfiguration Configuration { get; }
+        private IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
@@ -56,8 +56,7 @@ namespace OvgBlog.UI
             services.AddScoped<IUserService, UserService>();
             services.AddScoped<IContactService, ContactService>();
 
-
-            services.AddDbContextPool<OvgBlogContext>(x => x.UseSqlServer(Configuration["ConnectionStrings"].ToString()));           
+            services.AddDbContextPool<OvgBlogContext>(x => x.UseNpgsql(Configuration["ConnectionStrings"]?.ToString() ?? string.Empty));           
             
 
             services.AddTransient<IValidator<ArticleDetailViewModel>, ArticleDetailViewModelValidator>();
@@ -82,11 +81,11 @@ namespace OvgBlog.UI
                 options.Cookie.Name = "AUTHCOOKIE";
                 options.LoginPath = "/Login";
             });
+            
             services.AddRazorPages();
+            
         }
-
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {      
             if (env.IsDevelopment())
@@ -96,7 +95,6 @@ namespace OvgBlog.UI
             else
             {
                 app.UseExceptionHandler("/Error");
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
             //app.UseHttpsRedirection();
@@ -116,6 +114,13 @@ namespace OvgBlog.UI
 
                 endpoints.MapRazorPages();
             });
+            
+            using (var scope = app.ApplicationServices.CreateScope())
+            {
+                var db = scope.ServiceProvider.GetRequiredService<OvgBlogContext>();
+                db.Database.Migrate();
+                db.Seed(Configuration);
+            }
         }
     }
 }
